Move CacheableEntity to :build-cache-base
